MOSTRAR Actvdad y seson entero

+++ Codigo_GimnasioAdmin/src/negocio/actividad/TransActividad.java
@@ -6,12 +6,12 @@ public class TransActividad extends TransSesion {
 
 	private int aforo;
 
-	public TransActividad(int idM, int p, int a, String n) {
+	public TransActividad(int idM, double p, int a, String n) {
 		super(idM, p, n);
 		aforo = a;
 	}
 
-	public TransActividad(int i, int idM, int p, int a, String n) {
+	public TransActividad(int i, int idM, double p, int a, String n) {
 		super(i, idM, p, n);
 		aforo = a;
 	}
